organize code, put background color
now I am doing the next portion of the website with the star.

I went ahead and put the star as the background, placed the background-color

I was also able to figure out how to find the hex color through inspect

not I am moving the text content inside the star div. lol harder than I thought it would be

I cant really figure out how to put that space in between "Free coffee " and "is a tap  away" ugh