re-align the flexbox
lol i didnt think i would be kinda stuck in this section.

I was having trouble with the <p> tag in the sub-row1 b/c i wanted to make sure that it looked EXACTLY liked the one in the website

So what I did was change the width for .sub-Row1 to 80% so it centered the flexbox underneath the title

Then I got the p tag for sub class and I did a padding of left and right to 25px. i did not want to change the padding for the while flexbox. that't why i just changed the p tag

And it worked! yay im smart-ish