made learn.js for the Content-Torw 2 tabs
luckily the website that I found (frontend practice) provided some tips to help me make the website

It gave me a link to w3school on how to create tabs. when i click on a button it will display that particular image

amazingly enough i was able to figure out (through google of course) how to just put everything in the JS page b/c in w3school it showed the inline onclick and i want to avoid using any use in the inline.