created new folder to keep css
realized how unorganized my css code was so I placed it in different files and made a new folder to keep it all in to make my code clean.

i realize that if i wanted to use this css for the future i might as well keep it separated.

I also changed some of my index.html but tbh i dont really remember lol