holy moly
a lof of things happened here..

first of all in the index.html i changed the last two tabs class b/c it was inputted in corectly.

then for the LONGEST time i was trying to figure out how to make the transition when clicking on the tabs smooth

i was about to use  jQuery b/c that's all i saw when fading in and out the tabs

then i found out through the website that i found to make the website (frontend practice) that they had a link where i could create toggable tabs.

ugh it took me a while but honestly i just kept googling until i found the answer....

i dont even know if this will make me a developer v/c i keep trying to find the answer online

but t least i was able to finally to a smooth transition for the tabs..

i have to remember that (or not?)